# Conflicts:
#	src/KOKKOS/nbin_kokkos.cpp
#	src/KOKKOS/nbin_ssa_kokkos.cpp
#	src/KOKKOS/npair_kokkos.cpp

+++ src/KOKKOS/nbin_kokkos.cpp
@@ -156,7 +156,7 @@ void NBinKokkos<DeviceType>::binatomsItem(const int &i) const
   if (!Kokkos::isfinite(x(i, 0)) || !Kokkos::isfinite(x(i, 1)) || !Kokkos::isfinite(x(i, 2)))
     Kokkos::abort("Non-numeric positions - simulation unstable");
 
-  const int ibin = coord2bin(x(i, 0), x(i, 1), x(i, 2));
+  const int ibin = coord2bin(static_cast<double>(x(i, 0)), static_cast<double>(x(i, 1)), static_cast<double>(x(i, 2)));
 
   atom2bin(i) = ibin;
   const int ac = Kokkos::atomic_fetch_add(&bincount[ibin], (int)1);
