+++ src/KOKKOS/nbin_ssa_kokkos.cpp
@@ -242,7 +242,7 @@ void NBinSSAKokkos<DeviceType>::binIDAtomsItem(const int &i, int &update) const
     Kokkos::abort("Non-numeric positions - simulation unstable");
 
   int loc[3];
-  const int ibin = coord2bin(x(i, 0), x(i, 1), x(i, 2), &(loc[0]));
+  const int ibin = coord2bin(static_cast<double>(x(i, 0)), static_cast<double>(x(i, 1)), static_cast<double>(x(i, 2)), &(loc[0]));
   binID(i) = ibin;
 
   // Find the bounding box of the local atoms in the bins
@@ -269,7 +269,7 @@ void NBinSSAKokkos<DeviceType>::binIDGhostsItem(const int &i, int &update) const
     return;
   }
 
-  const int iAIR = coord2ssaAIR(x(i, 0), x(i, 1), x(i, 2));
+  const int iAIR = coord2ssaAIR(static_cast<double>(x(i, 0)), static_cast<double>(x(i, 1)), static_cast<double>(x(i, 2)));
   binID(i) = iAIR;
   if (iAIR > 0) { // include only ghost atoms in an AIR
     const int ac = Kokkos::atomic_fetch_add(&gbincount[iAIR], (int)1);
